refactor: global 패키지로 이동
- global 패키지에 도메인을 제외한 패키지를 이동시킨다.
