--- src/main/java/roomescape/auth/domain/LoginMember.java
@@ -1,6 +1,0 @@
-package roomescape.auth.domain;
-
-import roomescape.member.domain.MemberRole;
-
-public record LoginMember(Long id, String name, MemberRole memberRole) {
-}
