--- src/main/java/roomescape/auth/dto/CheckLoginResponse.java
@@ -1,10 +1,0 @@
-package roomescape.auth.dto;
-
-import roomescape.auth.domain.LoginMember;
-
-public record CheckLoginResponse(String name) {
-
-    public static CheckLoginResponse from(final LoginMember loginMember) {
-        return new CheckLoginResponse(loginMember.name());
-    }
-}
