--- src/main/java/roomescape/auth/dto/LoginRequest.java
@@ -1,4 +1,0 @@
-package roomescape.auth.dto;
-
-public record LoginRequest(String email, String password) {
-}
