--- src/main/java/roomescape/auth/dto/LoginResponse.java
@@ -1,4 +1,0 @@
-package roomescape.auth.dto;
-
-public record LoginResponse(String accessToken) {
-}
