--- src/main/java/roomescape/auth/exception/AuthException.java
@@ -1,8 +1,0 @@
-package roomescape.auth.exception;
-
-public class AuthException extends RuntimeException {
-
-    public AuthException(final String message) {
-        super(message);
-    }
-}
