--- src/main/java/roomescape/auth/exception/UnAuthorizedException.java
@@ -1,8 +1,0 @@
-package roomescape.auth.exception;
-
-public class UnAuthorizedException extends RuntimeException{
-
-    public UnAuthorizedException(final String message) {
-        super(message);
-    }
-}
