--- src/main/java/roomescape/common/GlobalExceptionHandler.java
@@ -1,58 +1,0 @@
-package roomescape.common;
-
-import java.time.format.DateTimeParseException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.auth.exception.AuthException;
-import roomescape.auth.exception.UnAuthorizedException;
-import roomescape.common.exception.AlreadyExistsException;
-import roomescape.common.exception.NotFoundException;
-import roomescape.common.exception.ValidationException;
-import roomescape.member.exception.MemberDuplicatedException;
-import roomescape.reservationtime.exception.ReservationTimeInUseException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(DateTimeParseException.class)
-    public ResponseEntity<String> handleDateTimeParseException(DateTimeParseException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<String> handleNotFoundException(NotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(ReservationTimeInUseException.class)
-    public ResponseEntity<String> handleReservationTimeInUseException(ReservationTimeInUseException e) {
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
-    }
-
-    @ExceptionHandler(AlreadyExistsException.class)
-    public ResponseEntity<String> handleAlreadyExistsException(AlreadyExistsException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(ValidationException.class)
-    public ResponseEntity<String> handleValidationException(ValidationException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(AuthException.class)
-    public ResponseEntity<String> handleAuthException(AuthException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
-    }
-
-    @ExceptionHandler(MemberDuplicatedException.class)
-    public ResponseEntity<String> handleMemberDuplicatedException(MemberDuplicatedException e) {
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
-    }
-
-    @ExceptionHandler(UnAuthorizedException.class)
-    public ResponseEntity<String> handleUnAuthorizedException(UnAuthorizedException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
-    }
-}
