--- src/main/java/roomescape/common/exception/AlreadyExistsException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class AlreadyExistsException extends RuntimeException {
-
-    public AlreadyExistsException(final String message) {
-        super(message);
-    }
-}
