--- src/main/java/roomescape/common/exception/DuplicatedException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class DuplicatedException extends RuntimeException {
-
-    public DuplicatedException(final String message) {
-        super(message);
-    }
-}
