--- src/main/java/roomescape/common/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(final String message) {
-        super(message);
-    }
-}
