--- src/main/java/roomescape/common/exception/ValidationException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class ValidationException extends RuntimeException {
-
-    public ValidationException(final String message) {
-        super(message);
-    }
-}
