+++ src/main/java/roomescape/global/auth/annotation/MemberPrincipal.java
@@ -1,0 +1,11 @@
+package roomescape.global.auth.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MemberPrincipal {
+}
