+++ src/main/java/roomescape/global/auth/domain/LoginMember.java
@@ -1,0 +1,6 @@
+package roomescape.global.auth.domain;
+
+import roomescape.member.domain.MemberRole;
+
+public record LoginMember(Long id, String name, MemberRole memberRole) {
+}
