+++ src/main/java/roomescape/global/auth/dto/LoginRequest.java
@@ -1,0 +1,4 @@
+package roomescape.global.auth.dto;
+
+public record LoginRequest(String email, String password) {
+}
