+++ src/main/java/roomescape/global/auth/dto/LoginResponse.java
@@ -1,0 +1,4 @@
+package roomescape.global.auth.dto;
+
+public record LoginResponse(String accessToken) {
+}
