+++ src/main/java/roomescape/global/auth/exception/AuthException.java
@@ -1,0 +1,8 @@
+package roomescape.global.auth.exception;
+
+public class AuthException extends RuntimeException {
+
+    public AuthException(final String message) {
+        super(message);
+    }
+}
