+++ src/main/java/roomescape/global/auth/exception/UnAuthorizedException.java
@@ -1,0 +1,8 @@
+package roomescape.global.auth.exception;
+
+public class UnAuthorizedException extends RuntimeException{
+
+    public UnAuthorizedException(final String message) {
+        super(message);
+    }
+}
