+++ src/main/java/roomescape/global/auth/infrastructure/JwtProvider.java
@@ -1,0 +1,48 @@
+package roomescape.global.auth.infrastructure;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import java.util.Date;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import roomescape.member.domain.Member;
+
+@Component
+public class JwtProvider {
+
+    @Value("${security.jwt.token.secret-key}")
+    private String secretKey;
+    @Value("${security.jwt.token.expire-length}")
+    private long validityInMilliseconds;
+
+    public String createToken(final Member member) {
+        Date now = new Date();
+        Date validity = new Date(now.getTime() + validityInMilliseconds);
+
+        return Jwts.builder()
+                .setSubject(member.getId().toString())
+                .claim("name", member.getName())
+                .claim("role", member.getMemberRole().name())
+                .setIssuedAt(now)
+                .setExpiration(validity)
+                .signWith(SignatureAlgorithm.HS256, secretKey)
+                .compact();
+    }
+
+    public Claims getAllClaimsFromToken(final String token) {
+        return Jwts.parser()
+                .setSigningKey(secretKey)
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public boolean isInvalidToken(final String token) {
+        try {
+            return getAllClaimsFromToken(token).getExpiration()
+                    .before(new Date());
+        } catch (IllegalArgumentException e) {
+            return false;
+        }
+    }
+}
