+++ src/main/java/roomescape/global/exception/AlreadyExistsException.java
@@ -1,0 +1,8 @@
+package roomescape.global.exception;
+
+public class AlreadyExistsException extends RuntimeException {
+
+    public AlreadyExistsException(final String message) {
+        super(message);
+    }
+}
