+++ src/main/java/roomescape/global/exception/DuplicatedException.java
@@ -1,0 +1,8 @@
+package roomescape.global.exception;
+
+public class DuplicatedException extends RuntimeException {
+
+    public DuplicatedException(final String message) {
+        super(message);
+    }
+}
