+++ src/main/java/roomescape/global/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package roomescape.global.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(final String message) {
+        super(message);
+    }
+}
