+++ src/main/java/roomescape/global/exception/ValidationException.java
@@ -1,0 +1,8 @@
+package roomescape.global.exception;
+
+public class ValidationException extends RuntimeException {
+
+    public ValidationException(final String message) {
+        super(message);
+    }
+}
