+++ src/main/java/roomescape/member/exception/MemberDuplicatedException.java
@@ -1,6 +1,6 @@
 package roomescape.member.exception;
 
-import roomescape.common.exception.DuplicatedException;
+import roomescape.global.exception.DuplicatedException;
 
 public class MemberDuplicatedException extends DuplicatedException {
