+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.annotation.MemberPrincipal;
-import roomescape.auth.domain.LoginMember;
+import roomescape.global.auth.annotation.MemberPrincipal;
+import roomescape.global.auth.domain.LoginMember;
 import roomescape.reservation.dto.request.AdminReservationCreateRequest;
 import roomescape.reservation.dto.request.ReservationCreateRequest;
 import roomescape.reservation.dto.response.ReservationResponse;
