+++ src/main/java/roomescape/reservation/exception/ReservationAlreadyExistsException.java
@@ -1,6 +1,6 @@
 package roomescape.reservation.exception;
 
-import roomescape.common.exception.AlreadyExistsException;
+import roomescape.global.exception.AlreadyExistsException;
 
 public class ReservationAlreadyExistsException extends AlreadyExistsException {
