+++ src/main/java/roomescape/reservationtime/controller/ReservationTimeController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.annotation.AdminPrincipal;
-import roomescape.auth.domain.LoginMember;
+import roomescape.global.auth.annotation.AdminPrincipal;
+import roomescape.global.auth.domain.LoginMember;
 import roomescape.reservationtime.dto.request.ReservationTimeCreateRequest;
 import roomescape.reservationtime.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservationtime.dto.response.ReservationTimeResponse;
