+++ src/main/java/roomescape/reservationtime/exception/ReservationTimeNotFoundException.java
@@ -1,6 +1,6 @@
 package roomescape.reservationtime.exception;
 
-import roomescape.common.exception.NotFoundException;
+import roomescape.global.exception.NotFoundException;
 
 public class ReservationTimeNotFoundException extends NotFoundException {
