+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.annotation.AdminPrincipal;
-import roomescape.auth.domain.LoginMember;
+import roomescape.global.auth.annotation.AdminPrincipal;
+import roomescape.global.auth.domain.LoginMember;
 import roomescape.theme.dto.request.ThemeCreateRequest;
 import roomescape.theme.dto.response.ThemeResponse;
 import roomescape.theme.service.ThemeService;
