+++ src/main/java/roomescape/theme/exception/InvalidThemeException.java
@@ -1,6 +1,6 @@
 package roomescape.theme.exception;
 
-import roomescape.common.exception.ValidationException;
+import roomescape.global.exception.ValidationException;
 
 public class InvalidThemeException extends ValidationException {
