+++ src/test/java/roomescape/MissionStepTest.java
@@ -21,8 +21,8 @@ import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
-import roomescape.auth.dto.CheckLoginResponse;
-import roomescape.auth.dto.LoginRequest;
+import roomescape.global.auth.dto.CheckLoginResponse;
+import roomescape.global.auth.dto.LoginRequest;
 import roomescape.member.dto.MemberResponse;
 import roomescape.member.dto.SignupRequest;
 import roomescape.reservation.controller.ReservationController;
